requirements.txt
uvicorn~=0.34.2
fastapi~=0.115.12
pydantic~=2.11.4
SQLAlchemy~=2.0.41
passlib~=1.7.4
alembic~=1.16.1
python-jose~=3.4.0
redis~=6.1.0
pydantic-settings~=2.9.1
